feat: allow skipping NPM tagging (#157)
* feat: allow skipping NPM tagging

Adds a `--skip-npm-tag` option that prevents NPM from tagging the release as either `latest` or `next`. Note that npm publish does not actually allow disabling the tagging, so instead a catch-all dummy tag (`ignore`) is used for the release. Also when publishing from a branch that is not master, the dummy tag is used automatically.

* remove dummy tag after publish

* add log notice about using dummy tag